# Optimize `SyntaxMap.commentRangeBeforeOffset(_:)`