By applying this, linting KeychainAccess by SwiftLint-0.5.1 is reduced from:
```
swiftlint  21.87s user 0.24s system 97% cpu 22.611 total
```
to:
```
swiftlint  17.06s user 0.15s system 97% cpu 17.607 total
```

On Instruments, `commentRangeBeforeOffset()` is reduced from 9931ms:
<img width="941" alt="screenshot 2015-12-14 20 19 37" src="https://cloud.githubusercontent.com/assets/33430/11780082/a1bd7144-a2a2-11e5-9779-17cd4c8df8a0.png">
 to 4148ms:
<img width="939" alt="screenshot 2015-12-14 20 19 48" src="https://cloud.githubusercontent.com/assets/33430/11780090/aeaa77b2-a2a2-11e5-96a3-88d6ea96d30c.png">
 on that test.